+++ crates/stackable-operator/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
 - Add functionality to convert LogLevel to an OPA log level ([#798])
 
 [#798]: https://github.com/stackabletech/operator-rs/pull/798